requirements.txt
Flask==1.0.2
PyJWT==1.7.1
SQLAlchemy==1.1.13
